+++ app/(user)/_components/HomeCarousel/caroselComponent.tsx
