+++ app/(user)/_components/HomeCarousel/fetchFeaturedProducts.ts
