+++ app/(user)/page.tsx
@@ -1,6 +1,6 @@
-import CategoryCardContainer from './homecomponents/categorycardcontainer';
-import HeroBanner from './homecomponents/herobanner';
-import HomepageCarosel from './caroselComponents/homepageCarosel';
+import CategoryCardContainer from './_components/categorycardcontainer';
+import HeroBanner from './_components/herobanner';
+import HomeCarousel from './_components/HomeCarousel/HomeCarousel';
 
 export default function Home() {
   return (
@@ -11,7 +11,7 @@ export default function Home() {
       </div>
       <div className="m-4">
         <h2 className="text-2xl font-bold text-left mb-4">Featured Products</h2>
-        <HomepageCarosel />
+        <HomeCarousel />
       </div>
     </>
   );
