+++ app/auth/error/page.tsx
@@ -1,4 +1,4 @@
-import { ErrorCard } from "@/app/components/auth/error-card"
+import { ErrorCard } from "@/app/_components/auth/error-card"
 
 const AuthErrorPage = () => {
     return (
