+++ app/auth/new-password/page.tsx
@@ -1,4 +1,4 @@
-import { NewPasswordForm } from "@/app/components/auth/new-password-form"
+import { NewPasswordForm } from "@/app/_components/auth/new-password-form"
 
 const NewPasswordPage = () => {
     return(
