+++ app/auth/new-verification/page.tsx
@@ -1,4 +1,4 @@
-import { NewVerificationForm } from "@/app/components/auth/new-verification-form"
+import { NewVerificationForm } from "@/app/_components/auth/new-verification-form"
 
 const NewVerificationPage = () => {
     return(
