+++ app/auth/register/page.tsx
@@ -1,5 +1,5 @@
 
-import { RegisterForm } from "@/app/components/auth/register-form"
+import { RegisterForm } from "@/app/_components/auth/register-form"
 
 const RegisterPage = () => {
     return (
